components/vector_query/src/requirements.txt
qdrant-client
fastembed
pyyaml
